refactor: add Pokedex tailwindcss classes
