--- src/pages/Pokemon/styles.ts
@@ -1,6 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-  display: flex;
-  width: 100vw;
-`;
